move api keywords to shared folder
