+++ apps/gp2-frontend/src/shared/api.ts
