+++ apps/gp2-frontend/src/shared/state.ts
